[minor] Avoid using the delete operator when possible
This adds an `_eventsCount` property to keep track of the number
of distinct events. The value of this property is checked when an
event needs to be removed. If it is 0, it means that there are no
other events, so a new empty object can be assigned to the
`_events` reference instead of using the delete operator.