Add license, iife wrapper, and cleanup exports
I've added the license to the top of the file, which helps it be used standalone, or when injected as a `<script>` directly from bower_modules.

Likewise, when used as a direct script, there are a few leaks into global context, the iife wrapper prevents that.

With the IIFE wrapper, I added an explicit addition to the context/window.  This is so the script can be used outside of CJS environment, such as direct include with bower.